+++ cli/test/models/datetime_converter_model.model.dart
@@ -1,4 +1,5 @@
 import 'package:dataforge_annotation/dataforge_annotation.dart';
+import 'package:collection/collection.dart';
 
 part 'datetime_converter_model.model.data.dart';