+++ cli/test/models/ignore_test.model.dart
@@ -1,4 +1,5 @@
 import 'package:dataforge_annotation/dataforge_annotation.dart';
+import 'package:collection/collection.dart';
 
 part 'ignore_test.model.data.dart';