+++ cli/test/models/map_types.model.dart
@@ -1,4 +1,5 @@
 import 'package:dataforge_annotation/dataforge_annotation.dart';
+import 'package:collection/collection.dart';
 
 part 'map_types.model.data.dart';